dec12/1: add elixir solution
less state than the perl solution; one thing i noticed when solving
part1 in perl, but didn't dwell on, was that it doesn't matter where you
add your movements since they are all summed together in the end (of
course, this was initially discovered through a typo, you know one of
those "why did that work?" situations that you don't want to question
too much. :-)). but it also meant that the accounting can be reduced to
a single distance variable and a variable for heading.