2023: dec01: Make problem 1 handle streams
I.e: Don't read the same file twice, assuming it's some kind of
stream, like a socket or a pipe. This means we can reuse the
solution in problem two by calling it, without having to store
the intermediate results on disk.

The solution is not the same, as we now do a first pass to make
the first digit be the first character of the string, and the last
digit be the last character. We then take the first and last char
of the string and concat them. If there is only one char in the
string, we just repeat it (since it is both first and last).