2025: dec01: problem 1, but in erlang
Not really integrated in the aoc runner, because Erlang is
how Erlang is. I mean, I could do it... Are you asking me to?
Maybe some rebar trick to make it convert a file like 1.erl
to a module like problem_1.erl, compile it and run it under
the hood?

Roadmap.

But for now:

  erlc problems.erl
  erl -noshell -run problems part 1 -s init stop

It was some time since i wrote erlang last, so probably not
as pretty as it could have been.